requirements.txt
tokenizers
transformers
datasets
torch
pygithub